% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Create_Obj.R
\name{Run_Create_Obj}
\alias{Run_Create_Obj}
\title{This function creates specified spatial transcriptomics data object for further personalized downstream analysis}
\usage{
Run_Create_Obj(gene.matrix, matched.data, obj.type, tech, ss.radius = 3000)
}
\arguments{
\item{gene.matrix}{Gene count matrix. This is usually obtained from 'Run_ST' function.}

\item{matched.data}{Data frame which contains spatial localization matched with gene count matrix. This is usually obtained from 'Run_loc_match' function.}

\item{obj.type}{Type of spatial transcriptomics object being created, can be 'Seurat', 'SpatialExperiment', or 'AnnData'.}

\item{tech}{Type of spatial transcriptomics sequencing technology, can be "Visium", "Slideseq", "Curio-seeker", or "Stereoseq".}

\item{ss.radius}{Optional. Radius for filtering SlideSeq or Curio-seeker spots for Seurat object. Default is 3000.}
}
\value{
Created spatial transcriptomics data object as required by 'obj.type'.
}
\description{
This function creates specified spatial transcriptomics data object for further personalized downstream analysis
}
\details{
Current mainstream analytic tools for downstream includes: Seurat, SpatialExperiment supported tools, and Squidpy.
This function can help creates corresponding objects for further downstream analysis.
}
\examples{
gene.matrix <- matrix(sample(1:100, 10 * 5, replace = TRUE), nrow = 10)
rownames(gene.matrix) <- paste0("Gene", 1:10)
colnames(gene.matrix) <- paste0("Spot", 1:5)
matched.data <- data.frame(
cell_name = paste0("Spot", 1:5),
barcode_sequence = paste0("BC", 1:5),
xcoord = runif(5, 0, 100),
ycoord = runif(5, 0, 100),
stringsAsFactors = FALSE)
obj <- Run_Create_Obj(gene.matrix = gene.matrix, 
matched.data = matched.data, 
obj.type = "SpatialExperiment", 
tech = "Visium")
}
