% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Vis.R
\name{Run_Vis}
\alias{Run_Vis}
\title{Visualize pre-processed sST data}
\usage{
Run_Vis(
  matched.data = NULL,
  config,
  Vis.spatial = TRUE,
  Vis.read = TRUE,
  show.config = TRUE
)
}
\arguments{
\item{matched.data}{A data frame containing spatial transcriptomics data, including UMI counts and spatial coordinates. This is usually obtained from 'Run_loc_match' function.}

\item{config}{Path to the YAML configuration file.}

\item{Vis.spatial}{Logical value indicating whether to visualize spatial data. Defaults to TRUE.}

\item{Vis.read}{Logical value indicating whether to visualize read-level data. Defaults to TRUE.}

\item{show.config}{Logical value indicating whether to print the configuration. Defaults to TRUE.}
}
\value{
A list contains spatial and read level visualization results.
}
\description{
This function visualizes both spatial-level and read-level information either before or after the 'Run_QC' step. It outputs raw and log-transformed UMI count plots for spatial flag, and demultiplexing and mapping statistics for read flag.
}
\examples{
\dontrun{
Vis.results <- Run_Vis(matched.data = matching, config = "/path/to/config_stPipe.yml", Vis.spatial = TRUE, Vis.read = TRUE, show.config = TRUE)
}
}
