% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Clustering.R
\name{Run_Clustering}
\alias{Run_Clustering}
\title{Run_Clustering: Perform Basic Clustering Algorithms}
\usage{
Run_Clustering(gene.count, matched.data)
}
\arguments{
\item{gene.count}{A data frame containing gene count data with gene IDs as row names.}

\item{matched.data}{A data frame containing matched spatial coordinates with raw UMI counts.}
}
\description{
This function performs basic clustering using t-SNE and UMAP, either before or after the QC step and outputs visualizations to output directory.
}
\examples{
clustering.result <- Run_Clustering(gene.count = gene_count, matched.data =  matching)
}
