% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_HTML.R
\name{Run_HTML}
\alias{Run_HTML}
\title{HTML report generation in RMarkDown format}
\usage{
Run_HTML(path)
}
\arguments{
\item{path}{Path to where results are stored, including the ones obtained from 'Run_ST', 'Run_loc_match', 'Run_Vis', and 'Run_Clustering'.}
}
\value{
Generated HTML report in RMarkDown format.
}
\description{
This function generates a HTML report of stPipe upstream processing steps in Rmarkdown format. It extracts plots in the result directory and outputs them in the whole report with corresponding explanation.
}
\examples{
\dontrun{
Run_HTML(path = "/path/to/ST_results")
}
}
