% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_loc_match.R
\name{Run_loc_match}
\alias{Run_loc_match}
\title{Match spatial location After Upstream Pre-Processing for Sequencing-Based Spatial Transcriptomics}
\usage{
Run_loc_match(config, show.config = TRUE)
}
\arguments{
\item{config}{Path to the YAML configuration file.}

\item{show.config}{Logical value indicating whether to print the configuration. Defaults to TRUE.}
}
\value{
A data frame contains filtered gene counts and spatial coordinates
}
\description{
This function matches spatial coordinates from sequencing-based spatial transcriptomics data after upstream pre-processing.
}
\examples{
matching <- Run_loc_match(config = "~/Desktop/config_stPipe.yml", show.config = TRUE)
}
