% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_QC.R
\name{Run_QC}
\alias{Run_QC}
\title{QC Control After Upstream Pre-Processing for Sequencing-Based Spatial Transcriptomics}
\usage{
Run_QC(config, matched.data, gene.matrix, show.config = TRUE)
}
\arguments{
\item{config}{Path to the YAML configuration file.}

\item{matched.data}{A data frame containing spatial transcriptomics data, including UMI counts and spatial coordinates, this is usually obtained from 'Run_loc_match' function.}

\item{gene.matrix}{A gene count matrix, this is usually obtained from 'Run_ST' function.}

\item{show.config}{Logical value indicating whether to print the configuration. Defaults to TRUE.}
}
\value{
A list containing filtered gene counts and spatial coordinates after QC.
}
\description{
This function performs QC control on sequencing-based spatial transcriptomics data after upstream pre-processing step such as 'Run_ST' step.
Filtering is performed either use specific UMI threshold or assign the threshold to 'DropletUtils'.
}
\examples{
qc.results <- Run_QC(config = "~/Desktop/config_stPipe.yml", matched.data = matching, gene.matrix = gene_count, show.config = TRUE)
}
