% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_ST.R
\name{Run_ST}
\alias{Run_ST}
\title{Pre-processing function for sequencing-based spatial transcriptomics}
\usage{
Run_ST(config, show.config = TRUE)
}
\arguments{
\item{config}{Path to the YAML configuration file.}

\item{show.config}{Logical value indicating whether to print the configuration. Defaults to TRUE.}
}
\value{
None. Outputs are saved to specified directories.
}
\description{
Pre-processing from FASTQ files to gene count matrix and extract spatial location information,
processing multiple sample is done via parallel computing. NB: for multiple samples, 'species' and 'technology_version' should be the same.
This function processes sequencing-based spatial transcriptomics data using various steps, including BAM to FASTQ conversion, trimming, index building, alignment, and barcode detection.
For Slideseq technology, the input should be BAM file and for all other technologies the input should be FASTQ file.
}
\examples{
\dontrun{
Run_ST(config = demo_config_path, show.config = TRUE)
}
}
