% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_ST.R
\name{Run_ST}
\alias{Run_ST}
\title{Run_ST: Pre-process from FASTQ files to gene count matrix and extract spatial location information}
\usage{
Run_ST(config, show.config = TRUE)
}
\arguments{
\item{config}{Path to the YAML configuration file.}

\item{show.config}{Logical value indicating whether to print the configuration. Defaults to TRUE.}
}
\description{
This function processes sequencing-based spatial transcriptomics data using various steps, including BAM to FASTQ conversion, trimming, index building, alignment, and barcode detection.
For Slideseq technology, the input should be BAM file and for all other technologies the input should be FASTQ file.
}
\examples{
mousebrain_visium <- Run_ST(config = "~/Desktop/config_stPipe.yml", show.config = TRUE)
}
