% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{stPipe}
\alias{stPipe}
\alias{stPipe-package}
\title{stPipe: A package for pre-processing sequencing-based spatial transcriptomics data.}
\description{
The stPipe will do spatial barcode demultiplexing, UMI deduplication, spatial location matching and quality control on
fastq data generated from all mainstream protocols
}
\section{stPipe functions}{

The stPipe functions Run_ST, Run_loc_match, Run_QC, Run_Vis, Run_Clustering, Run_create_obj, Run_HTML, Run_Interactive
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mritchielab/stPipe}
  \item Report bugs at \url{https://github.com/mritchielab/stPipe}
}

}
\author{
Yang Xu <xu.ya@wehi.edu.au>
}
